fix(build): resolve 100 compilation errors from second CI run
- Add QueryPerformanceAspect.QueryStats.getMinTime() stub
- Fix long→int cast for PasswordResetRequestedEvent.expiryMinutes
- Fix CSVFormat.Builder.get() → .build() (commons-csv 1.11.0 API)
- Remove non-existent OrderPlacedEvent.itemCount() and OrderCancelledEvent.reason() calls
- Fix Instant.format() in UserResolver — use formatter.format(instant.atZone(...))
- Fix all .createdAt(LocalDateTime.now()) builder calls on Instant fields
  (ActivityLogService x6, SecurityActivityService, AdminPromotionService,
   SellerPromotionService, SellerProductTagService)
- Add findByPublicId(UUID) to AdminPromotionRepository and SellerPromotionRepository;
  replace findById(UUID) calls and promotion.getId()→promotion.getPublicId()
- Change OrderRepository.findByCustomerId/findBySellerId/findSellerOrdersWithFilters/
  countSellerOrdersByStatusGrouped from Long to UUID (User.id is UUID)
- Change OrderItemRepository.countByProductSellerId/sumRevenueBySellerId Long→UUID
- Fix AnalyticsServiceImpl Instant/LocalDateTime comparison using atZone().toInstant()
- Add findByPublicId(UUID) to CartRepository; inject UserRepository into CartServiceImpl;
  replace Cart.builder().userId() with .user(user) lookup; fix cart.getUserId()→getUser().getId()
- Fix AbandonedCartScheduler cart.getUserId()→cart.getUser().getId()

+++ src/main/java/ecommerce/modules/auth/service/impl/AuthServiceImpl.java
@@ -369,7 +369,7 @@ public class AuthServiceImpl implements AuthService {
 
             eventPublisher.publish(new PasswordResetRequestedEvent(
                     user.getId(), user.getEmail(), user.getFirstName() + " " + user.getLastName(),
-                    token, tokenProperties.getPasswordResetMinutes()));
+                    token, (int) tokenProperties.getPasswordResetMinutes()));
 
             log.info("Password reset requested for user: {}", email);
         });
